made a zip mid files for fmri

+++ file_organization/unzip_raw_files.sh
@@ -12,11 +12,13 @@
 Subject_dir=$1
 remove_processed=$2
 
-cd ${Subject_dir}/Raw/MRI_files
-pwd
-unzip -n *.zip
+if [[ -d ${Subject_dir} ]]; then
+    cd ${Subject_dir}/Raw/MRI_files
+    pwd
+    unzip -n *.zip
 
-if [[ $remove_processed == "remove_proc" ]]; then
-    cd ${Subject_dir}/Processed
-    rm -r MRI_files
+    if [[ $remove_processed == "remove_proc" ]]; then
+        cd ${Subject_dir}/Processed
+        rm -r MRI_files
+    fi
 fi
